refactor(test): moved infrastructure-tied stubs out of test/domain/doubles
stub_webhook_submitter.go and stub_github_tokenizer.go double types that
live in internal/infrastructure/controllers/webhooks (Submitter, Job,
GitHubTokenizer) and therefore had to import that infrastructure package
to compile. Sitting under test/domain/doubles/repositories/ alongside
domain-contract stubs (StubAIReviewerRepository, StubRulesRepository,
etc.) made the domain doubles tree depend on infrastructure code.

Moved both stubs to test/infrastructure/doubles/repositories/, matching
the project layout documented in ~/.claude/rules/golang.md (which lists
both test/domain/doubles/repositories/ and test/infrastructure/doubles/
repositories/ as legitimate doubles roots).

Updated CLAUDE.md to record the convention so future stubs land on the
correct side of the import boundary.

Verified:
- go test -tags unit ./...     all packages OK
- make lint                    0 issues
