+++ internal/infrastructure/controllers/webhooks/github_test.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/rios0rios0/codeguru/internal/domain/entities"
 	"github.com/rios0rios0/codeguru/internal/infrastructure/controllers/webhooks"
-	doubles "github.com/rios0rios0/codeguru/test/domain/doubles/repositories"
+	doubles "github.com/rios0rios0/codeguru/test/infrastructure/doubles/repositories"
 )
 
 const (
